+++ src/main/java/dev/books/BookRegistration/Reviews/ReviewController.java
@@ -1,5 +1,8 @@
 package dev.books.BookRegistration.Reviews;
 
+import dev.books.BookRegistration.Books.BookDTO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,28 +20,45 @@ public class ReviewController {
     }
 
     @PostMapping("/create")
-    public ReviewDTO createReview(@RequestBody ReviewDTO review){
-        return reviewService.createReview(review);
+    public ResponseEntity<String> createReview(@RequestBody ReviewDTO review){
+        ReviewDTO newReview = reviewService.createReview(review);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Review criada com sucesso.");
     }
 
     @GetMapping("/list/{id}")
-    public ReviewDTO showReviewById(@PathVariable UUID id){
-        return reviewService.showReviewById(id);
+    public ResponseEntity<?> showReviewById(@PathVariable UUID id){
+        ReviewDTO review = reviewService.showReviewById(id);
+        if (review != null){
+            return ResponseEntity.ok(review);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("A review com ID " + id + " não foi encontrada.");
+        }
     }
 
     @GetMapping("/list")
-    public List<ReviewDTO> showAllReviews(){
-        return reviewService.showAllReviews();
+    public ResponseEntity<List<ReviewDTO>> showAllReviews(){
+        List<ReviewDTO> reviews = reviewService.showAllReviews();
+        return ResponseEntity.ok(reviews);
     }
 
     @PutMapping("/update/{id}")
-    public ReviewDTO updateReview(@PathVariable UUID id, @RequestBody ReviewDTO review){
-        return reviewService.updateReview(id, review);
+    public ResponseEntity<String> updateReview(@PathVariable UUID id, @RequestBody ReviewDTO review){
+       if (reviewService.showReviewById(id) != null){
+           reviewService.updateReview(id, review);
+           return ResponseEntity.ok("A review foi atualizada com sucesso.");
+       } else {
+           return ResponseEntity.status(HttpStatus.NOT_FOUND).body("A review com ID " + id + " não foi encontrada.");
+       }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteReview(@PathVariable UUID id){
-        reviewService.deleteReview(id);
+    public ResponseEntity<String> deleteReview(@PathVariable UUID id){
+        if (reviewService.showReviewById(id) != null){
+            reviewService.deleteReview(id);
+            return ResponseEntity.ok("Review foi removida com sucesso.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("A review com ID " + id + " não foi encontrado.");
+        }
     }
